Extract completion types and sorting into microcrate
