+++ Cargo.toml
@@ -42,6 +42,7 @@ members = [
     "crates/perl-lsp-feature-governance",
     "crates/perl-lsp-launcher",
     "crates/perl-lsp-navigation",
+    "crates/perl-lsp-completion-types",
     "crates/perl-lsp-completion",
     "crates/perl-lsp-code-actions",
     "crates/perl-lsp-cancellation",
@@ -137,6 +138,7 @@ allow = [
     "perl-lsp-diagnostic-types",
     "perl-lsp-diagnostics",
     "perl-lsp-code-actions",
+    "perl-lsp-completion-types",
     "perl-lsp-completion",
     "perl-lsp-inlay-hints",
 
@@ -292,6 +294,7 @@ perl-lsp-feature-governance = { path = "crates/perl-lsp-feature-governance", ver
 perl-lsp-launcher = { path = "crates/perl-lsp-launcher", version = "0.10.0" }
 perl-lsp-navigation = { path = "crates/perl-lsp-navigation", version = "0.10.0" }
 perl-lsp-rename = { path = "crates/perl-lsp-rename", version = "0.10.0" }
+perl-lsp-completion-types = { path = "crates/perl-lsp-completion-types", version = "0.10.0" }
 perl-lsp-completion = { path = "crates/perl-lsp-completion", version = "0.10.0" }
 perl-lsp-code-actions = { path = "crates/perl-lsp-code-actions", version = "0.10.0" }
 perl-diagnostics-codes = { path = "crates/perl-diagnostics-codes", version = "0.10.0" }
