+++ crates/perl-lsp-completion-types/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "perl-lsp-completion-types"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+description = "Shared completion item contracts and ordering policy for perl-lsp completion providers"
+readme = "README.md"
+license = "MIT OR Apache-2.0"
+repository.workspace = true
+homepage.workspace = true
+documentation = "https://docs.rs/perl-lsp-completion-types"
+keywords = ["perl", "lsp", "completion", "types", "ide"]
+categories = ["development-tools", "text-editors"]
+
+[lib]
+doctest = false
+
+[dependencies]
+perl-parser-core = { workspace = true }
+
+[lints]
+workspace = true
