+++ crates/perl-lsp-completion-types/README.md
@@ -1,0 +1,5 @@
+# perl-lsp-completion-types
+
+Shared completion item contracts and deterministic ordering policy for Perl LSP completion pipelines.
+
+This microcrate keeps data model and sorting behavior in one place so completion providers can stay focused on context extraction.
