+++ crates/perl-lsp-completion/src/completion/file_path.rs
@@ -38,7 +38,7 @@
 //! - **Symlink following**: Disabled for security
 
 use super::context::CompletionContext;
-use super::items::{CompletionItem, CompletionItemKind};
+use perl_lsp_completion_types::{CompletionItem, CompletionItemKind};
 #[cfg(not(target_arch = "wasm32"))]
 use std::path::{Component, Path, PathBuf};
