+++ crates/perl-lsp-completion/src/completion/packages.rs
@@ -2,7 +2,8 @@
 //!
 //! Provides completion for package members using workspace index integration.
 
-use super::{context::CompletionContext, items::CompletionItem};
+use super::CompletionContext;
+use perl_lsp_completion_types::CompletionItem;
 use perl_workspace_index::workspace_index::{SymbolKind as WsSymbolKind, WorkspaceIndex};
 use std::sync::Arc;
 
@@ -33,7 +34,7 @@ pub fn add_package_completions(
                 if symbol.name.starts_with(member_prefix) {
                     completions.push(CompletionItem {
                         label: symbol.name.clone(),
-                        kind: crate::completion::items::CompletionItemKind::Function,
+                        kind: perl_lsp_completion_types::CompletionItemKind::Function,
                         detail: Some(package_name.clone()),
                         documentation: symbol.documentation.clone(),
                         insert_text: Some(symbol.name.clone()),
@@ -48,7 +49,7 @@ pub fn add_package_completions(
                 if symbol.name.starts_with(member_prefix) {
                     completions.push(CompletionItem {
                         label: symbol.name.clone(),
-                        kind: crate::completion::items::CompletionItemKind::Variable,
+                        kind: perl_lsp_completion_types::CompletionItemKind::Variable,
                         detail: Some(package_name.clone()),
                         documentation: symbol.documentation.clone(),
                         insert_text: Some(symbol.name.clone()),
@@ -63,7 +64,7 @@ pub fn add_package_completions(
                 if symbol.name.starts_with(member_prefix) {
                     completions.push(CompletionItem {
                         label: symbol.name.clone(),
-                        kind: crate::completion::items::CompletionItemKind::Constant,
+                        kind: perl_lsp_completion_types::CompletionItemKind::Constant,
                         detail: Some(package_name.clone()),
                         documentation: symbol.documentation.clone(),
                         insert_text: Some(symbol.name.clone()),
